eldercare/eldercare/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.12.0
pydantic==2.5.3
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
azure-ai-inference==1.0.0b9
SpeechRecognition==3.10.0
pyaudio==0.2.13
pyttsx3==2.90
python-dateutil==2.8.2
asyncio==3.4.3 